pyproject.toml
[project]
name = "web_practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.8",
    "django-environ>=0.12.0",
    "kuro-py>=0.6.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
